Adjust nav bar
Previously, rails would throw an error when logged out because the nav made a reference to. Also, I had an previously unmerged commit (08099ab81e8d424ce2ece2f5fd1a920a87127b21) that is now merged with the team repo. However, I decided to ditch the work in that commit. Now, when logged out, the nav bar only includes the name Meter-Man, and not the login/logout links.